Trekker ut Stub for AaregRestClient som egen klasse
