+++ src/test/java/no/nav/fo/veilarbregistrering/arbeidsforhold/adapter/AaregGatewayTest.java
@@ -1,6 +1,5 @@
 package no.nav.fo.veilarbregistrering.arbeidsforhold.adapter;
 
-import no.nav.fo.veilarbregistrering.FileToJson;
 import no.nav.fo.veilarbregistrering.arbeidsforhold.AaregGateway;
 import no.nav.fo.veilarbregistrering.arbeidsforhold.Arbeidsforhold;
 import no.nav.fo.veilarbregistrering.arbeidsforhold.FlereArbeidsforhold;
@@ -16,7 +15,7 @@ public class AaregGatewayTest {
     private AaregGateway aaregGateway = new AaregGatewayImpl(new StubAaregRestClient());
 
     @Test
-    public void sak() {
+    public void hent_arbeidsforhold_fra_aareg_via_rest() {
         FlereArbeidsforhold flereArbeidsforhold = aaregGateway.hentArbeidsforhold(Foedselsnummer.of("12345678910"));
         assertThat(flereArbeidsforhold.siste()).isEqualTo(
                 new Arbeidsforhold()
@@ -24,17 +23,5 @@ public class AaregGatewayTest {
                         .setStyrk("2130123")
                         .setFom(LocalDate.of(2014, 7, 1))
                         .setTom(LocalDate.of(2015, 12, 31)));
-    }
-
-    class StubAaregRestClient extends AaregRestClient {
-
-        StubAaregRestClient() {
-            super(null, null);
-        }
-
-        @Override
-        protected String utforRequest(Foedselsnummer fnr) {
-            return FileToJson.toJson("/arbeidsforhold/arbeidsforhold.json");
-        }
     }
 }
