+++ src/test/java/no/nav/fo/veilarbregistrering/arbeidsforhold/adapter/AaregRestClientTest.java
@@ -1,6 +1,5 @@
 package no.nav.fo.veilarbregistrering.arbeidsforhold.adapter;
 
-import no.nav.fo.veilarbregistrering.FileToJson;
 import no.nav.fo.veilarbregistrering.bruker.Foedselsnummer;
 import org.junit.jupiter.api.Test;
 
@@ -38,17 +37,5 @@ public class AaregRestClientTest {
         arbeidsforholdDto.setArbeidsavtaler(Arrays.asList(arbeidsavtaleDto));
 
         assertThat(arbeidsforholdDtos).containsOnly(arbeidsforholdDto);
-    }
-
-    class StubAaregRestClient extends AaregRestClient {
-
-        StubAaregRestClient() {
-            super(null, null);
-        }
-
-        @Override
-        protected String utforRequest(Foedselsnummer fnr) {
-            return FileToJson.toJson("/arbeidsforhold/arbeidsforhold.json");
-        }
     }
 }
